Updated Drupal Core, Modified Gitignore to Remove Vendor, README
This is a rather large commit. It firstly updates Drupal core to 8.6.1 after testing on that branch.
It also removes the vendor directory from being actively tracked. This slows down git pushes and pulls and can always
be reacquired by running composer install. Finally, the README.md was added to explain these changes and how one might
use this repo in a production and dev environment.